refactor to incorporate new event broker directory
