+++ src/tavi/backend/model/tavi_project_model.py
@@ -1,7 +1,7 @@
 """Tavi Project."""
 
-from tavi.event_broker.event_broker import EventBroker
-from tavi.event_broker.event_type import Event
+from tavi.meta.event.event_broker import EventBroker
+from tavi.meta.event.event_interface import Event
 
 from tavi.backend.model.interface.tavi_project_interface import TaviProjectInterface
 from tavi.library.data.model_response import ModelResponse, ResponseCode
